Reorganize project
Ré-organisation des différents fichiers.
Ajout des fichiers __init__ dans les répertoires lorsque nécessaire.
Modification des chemin d'import au début des différents scripts.
