+++ src/tests/test_model.py
@@ -1,0 +1,13 @@
+import sys
+print("path test model : ", sys.path)
+
+from models.model import mse_lgb, mae_lgb, r2_lgb, rmse_lgb
+
+import pytest
+
+# Vérification de la performance du modèle LGBM
+def test_r2():
+  assert r2_lgb > 0.5
+
+def test_rmse():
+  assert rmse_lgb < 120
